Former-commit-id: 5c65b042852d12adff8f33456fa9701245c999c0 [formerly 68d32fa1a2b9d47ae2f226b1701f7b23a36e03f8]
Former-commit-id: 832a75665f5df799236383b94e0de5cd2cf41847

+++ block/blockchain.js
@@ -50,8 +50,8 @@ blockchain.prototype.blockFromBytes = function (buffer) {
     b.timestamp = bb.readInt();
     b.previousBlock = bb.readLong();
     b.numbersOfTransactions = bb.readInt();
-    b.totalAmount = bb.readFloat();
-    b.totalFee = bb.readFloat();
+    b.totalAmount = bb.readFloat64();
+    b.totalFee = bb.readFloat64();
     b.payloadLength = bb.readInt();
 
     var payloadHash = new Buffer(32);
@@ -104,8 +104,8 @@ blockchain.prototype.pushBlock = function (buffer) {
     b.previousBlock = bignum.fromBuffer(pb).toString();
     b.numberOfAddresses = bb.readInt();
     b.numberOfTransactions = bb.readInt();
-    b.totalAmount = bb.readFloat();
-    b.totalFee = bb.readFloat();
+    b.totalAmount = bb.readFloat64();
+    b.totalFee = bb.readFloat64();
     b.payloadLength = bb.readInt();
 
     var payloadHash = new Buffer(32);
@@ -375,8 +375,6 @@ module.exports.addGenesisBlock = function (app, cb) {
         bc.lastBlock = b.getId();
         t.blockId = genesisblock.blockId;
 
-        //b.sign("nY4NxXNd9velmtPxRN6TS8JLDR2dMGzkyL51p1sTPefA3tY9SzWBZT6GYlxyUgCQhSrJsoLiXHiuGqFVZTEObqI5BWgua6i5MAk");
-        //console.log(b.blockSignature.toString('hex'), b.getId());
         b.setApp(app);
 
         var r = b.analyze();
